refactor api controllers
